# Optimize JSON encoding of numbers