This further optimizes the JSON encoding of integers and floats.